requirements.txt
flask
flask-bcrypt
flask-mail
python-dotenv
gunicorn
